+++ projects/sango/generator/composite/index.ts
@@ -1,40 +1,25 @@
-import { readdir, stat } from "../../fs_promise"
-import { FileLoader, FileLoaderParams } from "./FileLoader"
+import { FileLoaderParams } from "./FileLoader"
+import { loadFiles } from "./loadFiles"
 
 export interface CompositeFile {
   toYaml: string
   location: string
 }
 
-const setupLoader = async (params: FileLoaderParams) => {
-  const stats = await stat(params.path.raw)
-  return stats.isDirectory() ?
-    new DirectoryLoader(params) :
-    new FileLoader(params)
+export interface Composite {
+  parentFile: CompositeFile
+  childFiles: CompositeFile[]
 }
 
-class DirectoryLoader {
-  constructor (private readonly params: FileLoaderParams) { }
-  path = this.params.path
-
-  async run (): Promise<CompositeFile[]> {
-    const files = await readdir(this.path.raw)
-    const promises = files.map(this.toComposite)
-    const composites: CompositeFile[][] = await Promise.all(promises)
-    return composites.reduce((as, bs) => as.concat(bs), [])
-  }
-
-  private toComposite = async (file: string): Promise<CompositeFile[]> => {
-    return loadFiles({
-      discriminator: this.params.discriminator,
-      parent: this.params.parent,
-      path: this.path.append(file),
-    })
-  }
-}
-
-export const loadFiles =
-  async (params: FileLoaderParams): Promise<CompositeFile[]> => {
-    const loader = await setupLoader(params)
-    return loader.run()
+export const composeFiles =
+  async (params: FileLoaderParams): Promise<Composite> => {
+    const childFiles = await loadFiles(params)
+    const parentFile = {
+      location: "",
+      toYaml: "",
+    }
+    return {
+      parentFile,
+      childFiles,
+    }
   }
